moved ObservableLists to a new subpackage
